+++ docs/api/version-history.md
@@ -26,6 +26,7 @@ keywords: "API, Docker, rcli, REST, documentation"
   `BindOptions.ReadOnlyNonRecursive` and `BindOptions.ReadOnlyForceRecursive` to customize the behavior.
 * `POST /containers/create` now accepts a `HealthConfig.StartInterval` to set the
   interval for health checks during the start period.
+* A new `MacAddress` field has been added to `EndpointSettings` in API v1.44.
 
 ## v1.43 API changes